tried to fix linux

+++ src/Plugin.cpp
@@ -122,7 +122,7 @@ jobjectArray CallGlobal(JNIEnv* jenv, jclass jcl, jstring packageName, jstring f
 	size_t returnsLength = luaReturns.size();
 	jclass objectCls = jenv->FindClass("Ljava/lang/Object;");
 	jobjectArray returns = jenv->NewObjectArray((jsize)returnsLength, objectCls, NULL);
-	for (jsize i = 0; i < returnsLength; i++) {
+	for (jsize i = 0; i < (int) returnsLength; i++) {
 		jenv->SetObjectArrayElement(returns, i, env->ToJavaObject(Plugin::Get()->GetPackageState(packageNameStr), luaReturns[i]));
 	}
 	jenv->DeleteLocalRef(packageName);
